refactor: get candidates func to tree impl
